feat(cli): guard against missing native binaries
stash loads native Rust addons (protect-ffi via @cipherstash/stack, and
@cipherstash/auth) distributed as per-platform optional npm packages. npm
intermittently skips installing these optional deps (npm/cli#4828), leaving
the base package present but the platform binary missing. The failure was a
raw MODULE_NOT_FOUND stack trace with no guidance.

- Split the bin into a thin launcher (stash.ts) that dynamically imports the
  command graph (main.ts), so a native-load failure during module evaluation
  is caught and rendered as actionable fix guidance instead of a stack trace.
- Add src/native.ts: isNativeBinaryMissing() detection + reportNativeBinaryMissing()
  guidance, with unit tests.
- Add 'stash doctor' to diagnose runtime + native modules. Dispatched before
  the command graph loads so it runs even when a binary is missing.

Interim hardening ahead of the planned Rust/Homebrew CLI.
